# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: National Energy War Room — REST API
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /api/v1/sector/kpis:
    get:
      operationId: getSectorKpis
      tags: [sector]
      security:
        - bearerAuth: []
      summary: Get sector KPIs (staff only)
      responses:
        "200":
          description: KPI data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SectorKpis"
...
  /api/v1/sector/grid/stream:
    get:
      operationId: streamGrid
      tags: [sector]
      security:
        - bearerAuth: []
      summary: SSE stream of live grid metrics
      responses:
        "200":
          description: SSE stream
  /api/v1/sector/sankey:
    get:
      operationId: getSankeyData
      tags: [sector]
      security:
        - bearerAuth: []
      summary: Sankey diagram data (physical + financial flows)
      responses:
        "200":
          description: Sankey flow data
  /api/v1/sector/plants:
    get:
      operationId: getPlants
      tags: [sector]
      security:
        - bearerAuth: []
      summary: List generation plants
      parameters:
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Plants list
  /api/v1/sector/plants/{id}:
    get:
      operationId: getPlant
      tags: [sector]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plant detail
  /api/v1/sector/plants/{id}/units:
    get:
      operationId: getPlantUnits
      tags: [sector]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Plant units
  /api/v1/sector/transmission/lines:
    get:
      operationId: getTransmissionLines
      tags: [sector]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/page"
      responses:
        "200":
          description: Transmission lines
  /api/v1/sector/transmission/substations:
    get:
      operationId: getSubstations
      tags: [sector]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Substations
  /api/v1/sector/discos:
    get:
      operationId: getDiscos
      tags: [sector]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/page"
        - name: sort
          in: query
          schema:
            type: string
      responses:
        "200":
          description: DisCo list
  /api/v1/sector/discos/{id}:
    get:
      operationId: getDisco
      tags: [sector]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: DisCo detail
  /api/v1/sector/discos/{id}/feeders:
    get:
      operationId: getDiscoFeeders
      tags: [sector]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Feeder list
  /api/v1/sector/gas/pipelines:
    get:
      operationId: getGasPipelines
      tags: [sector]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Gas pipelines
  /api/v1/sector/gas/diversions:
    get:
      operationId: getDiversionOpportunities
      tags: [sector]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Diversion opportunities
  /api/v1/sector/projects:
    get:
      operationId: getCapitalProjects
      tags: [sector]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Capital projects
  /api/v1/sector/minigrids:
    get:
      operationId: getMiniGrids
      tags: [sector]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Mini-grids
  /api/v1/sector/settlement/invoices:
    get:
      operationId: getSettlementInvoices
      tags: [sector]
      security:
        - bearerAuth: []
      parameters:
        - name: period
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Settlement invoices
  /api/v1/sector/value-chain:
    get:
      operationId: getValueChain
      tags: [sector]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Value chain links
  /api/v1/sector/value-chain/{key}:
    get:
      operationId: getValueChainLink
      tags: [sector]
      security:
        - bearerAuth: []
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Value chain link detail
  /api/v1/sector/alerts/active:
    get:
      operationId: getActiveAlerts
      tags: [sector]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Active alerts
  /api/v1/sector/alerts/stream:
    get:
      operationId: streamAlerts
      tags: [sector]
      security:
        - bearerAuth: []
      summary: SSE stream of active alerts
      responses:
        "200":
          description: SSE stream
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    pageSize:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    SectorKpis:
      type: object
      properties:
        data:
          type: object
          properties:
            availableGenerationMw:
              type: number
            installedCapacityMw:
              type: number
            hourlyGenerationMwh:
              type: number
            atccLossPct:
              type: number
            meteringRatePct:
              type: number
            plantAvailabilityFactor:
              type: number
            openComplaints:
              type: integer
            frequencyHz:
              type: number
            source:
              type: string
...